some restructing
