+++ lib/drink_category/widgets/widgets.dart
@@ -1,2 +1,1 @@
-export 'drink_type_extension.dart';
 export 'drinks_list_widget.dart';
